requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-anthropic==0.1.11
langchain-community==0.0.36
chromadb==0.4.24
pdfplumber==0.11.0
Pillow==10.3.0
reportlab==4.1.0
python-dotenv==1.0.1
anthropic>=0.40.0
httpx>=0.27.0
pdf2image==1.17.0
pandas==2.2.1
plotly==5.20.0
streamlit-extras==0.4.2
